lib/atom.c: move here remaining code from atom-private.c
