+++ src/lib/Makefile.am
@@ -10,7 +10,6 @@ libfixedpoint_la_SOURCES = fixedpoint.h fixedpoint.c
 
 liblldpctl_la_SOURCES = \
 	lldpctl.h atom.h errors.c connection.c atom.c helpers.c helpers.h \
-	atom-private.c \
 	atoms/config.c atoms/dot1.c atoms/dot3.c \
 	atoms/interface.c atoms/med.c atoms/mgmt.c atoms/port.c
 liblldpctl_la_LIBADD  = $(top_builddir)/src/libcommon-daemon-lib.la libfixedpoint.la
